third_party/blink/web_tests/media/media-fragments/TC0072-expected.txt


Title: TC0072
Fragment: 't=3.'
Comment: Equivalent to #t=a,e. The media is requested from a to e.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
